Moved exploratory code to dedicated folder
That stuff is partly deprecated, anyway.
